test: Add unit tests for application layer use cases
- Implements a new test project 'MotorcycleRental.Application.UnitTests'.

- Adds unit tests for all 10 command and query handlers in the Application layer.

- Uses xUnit, NSubstitute, and FluentAssertions.

- Fixes several bugs in command handlers that were discovered during test creation.

+++ MotorcycleRental/src/Core/MotorcycleRental.Application/UseCases/DeliveryDrivers/UpdateCnhImage/UpdateCnhImageCommandHandler.cs
@@ -1,15 +1,9 @@
-﻿using MediatR;
+using MediatR;
 using MotorcycleRental.Application.Interfaces;
-using MotorcycleRental.Application.UseCases.DeliveryDrivers.CreateDeliveryDriver;
 using MotorcycleRental.Domain.Exceptions;
 using MotorcycleRental.Domain.Interfaces;
 using MotorcycleRental.Domain.Interfaces.Repositories;
 using MotorcycleRental.Domain.Validators;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace MotorcycleRental.Application.UseCases.DeliveryDrivers.UpdateCnhImage
 {
@@ -18,6 +12,7 @@ namespace MotorcycleRental.Application.UseCases.DeliveryDrivers.UpdateCnhImage
         private readonly IDeliveryDriverRepository _repository;
         private readonly IUnitOfWork _unitOfWork;
         private readonly IStorageService _storageService;
+
         public UpdateCnhImageCommandHandler(IDeliveryDriverRepository repository, 
             IUnitOfWork unitOfWork, IStorageService storageService)
         {
@@ -28,12 +23,19 @@ namespace MotorcycleRental.Application.UseCases.DeliveryDrivers.UpdateCnhImage
 
         public async Task Handle(UpdateCnhImageCommand request, CancellationToken cancellationToken)
         {
-            var isValid = ImageFormatValidator.IsValidImageFormat(request.CnhImage, out string? errorMessage);
+            ImageFormatValidator.IsValidImageFormat(request.CnhImage, out string? errorMessage);
 
             if (!string.IsNullOrEmpty(errorMessage))
+            {
                 throw new DomainException(errorMessage);
+            }
 
             var deliveryDriver = await _repository.GetByIdAsync(request.Id);
+
+            if (deliveryDriver is null)
+            {
+                throw new NotFoundException("Entregador não encontrado"); // Corresponds to a 404 Not Found
+            }
 
             await _storageService.DeleteAsync(deliveryDriver.Cnh.ImagePath);
             var path = await _storageService.SaveBase64FileAsync("cnh", request.CnhImage, "cnh_images");
